Refactor: change cmd name and remove unnecessary arguments
Refactor the `refactor` command to remove the `[fileout]` argument and change the cmd name to better reflect the remaining argument. This commit includes no breaking changes.

feat(cli): add new `baseurl` flag to set the API URL

In this feature, a new `baseurl` flag is added to the CLI command, allowing users to set a custom URL for the API. This feature does not include any breaking changes and is considered a minor update.

fix(cli): fix bug in `IndexFilesAndEmbeddings`

A bug was found in the `IndexFilesAndEmbeddings` function where the workflow was not properly returned when the `disableIndex` flag was set. This commit fixes that bug, without introducing any breaking changes.

chore: cleanup unused code in `runRefactorWithConfigFile`

Some code within the `runRefactorWithConfigFile` function was no longer being used, and has thus been removed. This commit includes no breaking changes.