refactor!: update OpenAI client initialization to use config values
This commit updates the initialization of the OpenAI client in the tzapconnect package to use configuration values from the config package. The initialization functions `InitiateOpenaiClient` and `initializeTzap` in the openaiconnector and root command files respectively now only require the OpenAI API key as input. The `EmbeddingURL` and `CompletionURL` values from the `tzapCliSettings` struct are now used to construct the clients, instead of the `BaseURL` value which has been removed. This improves the flexibility of the code, allowing the OpenAI client to be configured using various URLs for different services. No breaking changes have been made.