fix(cli): fix localwalker file traversal issue
The localwalker package in the cli/cmd/cmdinstance/localwalker/localwalker.go file had an issue with file traversal. When checking if a directory should be traversed or skipped, the code mistakenly used os.ReadDir instead of os.Stat. This resulted in skipping directories that should have been traversed. This commit fixes the issue by replacing os.ReadDir with os.Stat when checking if a directory should be traversed.

BREAKING CHANGE: The behavior of file traversal in the localwalker package has changed. Previously, directories were skipped when they should have been traversed. After this fix, directories will be traversed correctly. This may affect applications that rely on the previous behavior.

Closes #12